+++ changelog.md
@@ -7,6 +7,7 @@ vX.X.X
 - Fixed bug where "Local TX echo" does not work, closes #112.
 - Added a "New Tab" button/tab to the end of row of tab headers, closes #104.
 - Fixed bug where COM port was not closed if terminal tab was closed, closes #113.
+- Fixed bug where COM port and it's thread were not closed if COM port was open when the application exits, closes #114.
 
 v0.6.1
 ------